Correct the structure of main functions
