+++ modules/run_dna_rna_tools.py
@@ -108,35 +108,3 @@ def reverse_complement(seq: str)->str:
         - str: reversed complemetary DNA or RNA sequence
         """
     return reverse(complement(seq))
-
-
-def run_dna_rna_tools(*args)->list:
-    """
-    Process DNA- and RNA-sequence(s)
-    arguments:
-        - args: sequence or sequences and operation for it or them
-    operations:
-        -'transcribe': transcribe the DNA-sequence(s)
-        -'reverse': reverse the sequence(s)
-        -'complement': return the complemetary sequence(s)
-        -'reverse_complement': return the reversed complemetary sequence(s)
-    return:
-        - list (or str): processed sequence(s)
-    """
-    *seqs, procedure = args
-    res = []
-    functions = {
-        'transcribe': transcribe,
-        'reverse': reverse,
-        'complement': complement,
-        'reverse_complement': reverse_complement
-        }
-    for seq in seqs:
-        if is_dna(seq) or is_rna(seq):
-            after = functions[procedure](seq)
-            res.append(after)
-        else:
-            raise ValueError('Incorrect input sequence(s), please try again')
-    if len(res) == 1:
-        res = res[0]
-    return res
